backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Backend for the Rue take-home assignment"
authors = ["Alfredo Delgado <alfredodlgdmrn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
djangorestframework-simplejwt = "^5.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
